refactoring.....

+++ src/tools/path.py
@@ -1,50 +1,25 @@
 # -*- coding: utf-8 -*-
-
 import os
 import shutil
+import locale
 
 
 class Path(object):
     u"""
-
+    定义资源,生成的文件等的路径,以及关于路径操作的一些函数
     """
-    try:
-        base_path = unicode(os.path.abspath('.').decode('gbk'))  # 初始地址,不含分隔符
-    except:
-        base_path = os.path.abspath('.')  # 对于Mac和Linux用户，使用gbk解码反而会造成崩溃，故添加一个try-except，以防万一
-
+    base_path = unicode(os.path.abspath('.').decode(locale.getpreferredencoding()))
+    
     config_path = base_path + u'/jianshu_config.json'
     db_path = base_path + u'/jianshu_db_002.sqlite'
     sql_path = base_path + u'/db/jianshu.sql'
 
     www_css = base_path + u'/www/css'
-    www_image = base_path + u'/www/image'
+    www_image = base_path + u'/www/images'
 
-    html_pool_path = base_path + u'/电子书临时资源库/简书/网页池'
-    image_pool_path = base_path + u'/电子书临时资源库/简书/图片池'
-    result_path = base_path + u'/生成的电子书/简书'
-
-    @staticmethod
-    def mkdir(path):
-        try:
-            os.mkdir(path)
-        except OSError:
-            pass
-        return
-
-    @staticmethod
-    def chdir(path):
-        u"""
-        换路径,如果路径不存在就新建一个
-        :param path:
-        :return:
-        """
-        try:
-            os.chdir(path)
-        except OSError:
-            Path.mkdir(path)
-            os.chdir(path)
-        return
+    html_pool_path = base_path + u'/电子书临时资源库/网页池'
+    image_pool_path = base_path + u'/电子书临时资源库/图片池'
+    result_path = base_path + u'/生成的电子书'
 
     @staticmethod
     def reset_path():
@@ -58,17 +33,39 @@ class Path(object):
         :return:
         """
         print os.path.realpath('.')
+        return
 
     @staticmethod
     def get_pwd():
         u"""
         :return: 绝对路径
         """
-        try:
-            path = unicode(os.path.abspath('.').decode('gbk'))  # 初始地址,不含分隔符
-        except:
-            path = os.path.abspath('.')  # 对于Mac和Linux用户，使用gbk解码反而会造成崩溃，故添加一个try-except，以防万一
+        path = unicode(os.path.abspath('.').decode(locale.getpreferredencoding()))
         return path
+
+    @staticmethod
+    def mkdir(path):
+        try:
+            os.mkdir(path)
+        except OSError:
+            # Debug.logger.debug(u'指定目录已存在')
+            pass
+        return
+
+    @staticmethod
+    def chdir(path):
+        u"""
+        换路径,如果路径不存在就新建一个
+        :param path:
+        :return:
+        """
+        try:
+            os.chdir(path)
+        except OSError:
+            # Debug.logger.debug(u'指定目录不存在，自动创建之')
+            Path.mkdir(path)
+            os.chdir(path)
+        return
 
     @staticmethod
     def rmdir(path):
@@ -77,7 +74,9 @@ class Path(object):
         :param path:
         :return:
         """
-        shutil.rmtree(path, ignore_errors=True)
+        if path:
+            shutil.rmtree(path, ignore_errors=True)
+        return
 
     @staticmethod
     def copy(src, dst):
@@ -100,21 +99,18 @@ class Path(object):
         初始化路径,不需要实例化 Path 就能执行
         :return:
         """
-        try:
-            base_path = unicode(os.path.abspath('.').decode('gbk'))  # 初始地址,不含分隔符
-        except:
-            base_path = os.path.abspath('.')  # 对于Mac和Linux用户，使用gbk解码反而会造成崩溃，故添加一个try-except，以防万一
+        Path.base_path = Path.get_pwd()
 
-        Path.config_path = base_path + u'/jianshu_config.json'
-        Path.db_path = base_path + u'/jianshu_db_002.sqlite'
-        Path.sql_path = base_path + u'/db/jianshu.sql'
+        Path.www_css = Path.base_path + u'/www/css'
+        Path.www_image = Path.base_path + u'/www/images'
 
-        Path.www_css = base_path + u'/www/css'
-        Path.www_image = base_path + u'/www/image'
+        Path.config_path = Path.base_path + u'/jianshu_config.json'
+        Path.db_path = Path.base_path + u'/db/jianshu_db_002.sqlite'
+        Path.sql_path = Path.base_path + u'/db/jianshu.sql'
 
-        Path.html_pool_path = base_path + u'/电子书临时资源库/简书/网页池'
-        Path.image_pool_path = base_path + u'/电子书临时资源库/简书/图片池'
-        Path.result_path = base_path + u'/生成的电子书/简书'
+        Path.html_pool_path = Path.base_path + u'/电子书临时资源库/网页池'
+        Path.image_pool_path = Path.base_path + u'/电子书临时资源库/图片池'
+        Path.result_path = Path.base_path + u'/生成的电子书'
         return
 
     @staticmethod
@@ -122,11 +118,7 @@ class Path(object):
         Path.reset_path()
         Path.mkdir(u'./电子书临时资源库')
         Path.mkdir(u'./生成的电子书')
-        Path.chdir(u'./生成的电子书')
-        Path.mkdir(u'./简书')
-        Path.chdir(u'../电子书临时资源库')
-        Path.mkdir(u'./简书')
-        Path.chdir(u'./简书')
+        Path.chdir(u'./电子书临时资源库')
         Path.mkdir(u'./网页池')
         Path.mkdir(u'./图片池')
         Path.reset_path()
